Support ES384 Certificates (#33)
Add the es384-subroot circuit and refactor the es256-* circuits along with their witness generation logic.
This is still a very early version and will require significant optimization to be practical. For example, generating a proof for a single ES384 certificate on a Google Pixel 9a currently takes over 40 seconds.